requirements.txt
hydra-core
lightning
mido
numpy
omegaconf
sox
toml
torch
torchaudio
torchmetrics
tqdm
